# Add flag to skip optimization