Pretty small change but makes the compilation faster if you want a faster iteration cycle. 

Without skip-opt:

```
[I] ben@dylibso ~/d/js-pdk (main)> time ./target/release/extism-js examples/simple_js/script.js -i examples/simple_js/script.d.ts -o examples/simple_js.wasm
[2024-10-16T15:48:41Z INFO  extism_js] parsed interface: 298.5µs
[2024-10-16T15:48:41Z INFO  extism_js] read user code: 38.417µs
[2024-10-16T15:48:41Z INFO  extism_js] concated user code: 3.209µs
[2024-10-16T15:48:41Z INFO  extism_js] command done: 60.358042ms
[2024-10-16T15:48:41Z INFO  extism_js] generated shims: 267.958µs
[2024-10-16T15:48:41Z INFO  extism_js] wasm merge check: 22.89325ms
[2024-10-16T15:48:41Z INFO  extism_js] wasm merge: 125.382209ms
[2024-10-16T15:48:42Z INFO  extism_js] wasm opt: 774.119041ms

________________________________________________________
Executed in  995.36 millis    fish           external
   usr time    4.16 secs      0.24 millis    4.16 secs
   sys time    0.08 secs      1.16 millis    0.08 secs

```

With skip-opt:

```
[I] ben@dylibso ~/d/js-pdk (main)> time ./target/release/extism-js examples/simple_js/script.js -i examples/simple_js/script.d.ts -o examples/simple_js.wasm --skip-opt
[2024-10-16T15:48:44Z INFO  extism_js] parsed interface: 298.333µs
[2024-10-16T15:48:44Z INFO  extism_js] read user code: 37.458µs
[2024-10-16T15:48:44Z INFO  extism_js] concated user code: 2.625µs
[2024-10-16T15:48:44Z INFO  extism_js] command done: 59.516458ms
[2024-10-16T15:48:44Z INFO  extism_js] generated shims: 245.542µs
[2024-10-16T15:48:44Z INFO  extism_js] wasm merge check: 23.005541ms
[2024-10-16T15:48:44Z INFO  extism_js] wasm merge: 125.184334ms

________________________________________________________
Executed in  219.40 millis    fish           external
   usr time  280.54 millis  175.00 micros  280.37 millis
   sys time   36.47 millis  812.00 micros   35.66 millis
```